Removed TailwindCSS since we're not using it right now. Also removed Tailwind.macro. Completed refactor of countries into Stat components.
