+++ package.json
@@ -31,9 +31,7 @@
     "babel-plugin-macros": "^2.8.0",
     "babel-preset-gatsby": "^0.2.28",
     "gh-pages": "^2.1.1",
-    "prettier": "^1.18.2",
-    "tailwind.macro": "^1.0.0-alpha.10",
-    "tailwindcss": "^1.2.0-canary.5"
+    "prettier": "^1.18.2"
   },
   "repository": {
     "type": "git",
