+++ postcss.config.js
@@ -1,3 +1,2 @@
 module.exports = () => ({
-  plugins: [require("tailwindcss")("./tailwind.config.js")],
 })
