+++ src/styles/global.css
@@ -14,12 +14,6 @@
     url("/static/GT-America-Standard-Bold.otf") format("opentype");
 }
 
-@tailwind base;
-
-@tailwind components;
-
-@tailwind utilities;
-
 /* Color Variables */
 :root {
   --headline-color: #291a42;
